add: types directory for game
